+++ src/protocol/mod.rs
@@ -32,6 +32,7 @@ mod model;
 mod pilot;
 mod request;
 mod response;
+mod scene;
 mod types;
 
 pub use config::{ModelConfig, Power, SystemConfig, UserConfig};
@@ -41,4 +42,5 @@ pub use model::{
 pub use pilot::{Pilot, PilotBuilder, PilotParams, Success};
 pub use request::Request;
 pub use response::{DeviceError, Response};
+pub use scene::{Adjustable, Scene};
 pub use types::{Channel, Devices, Dimming, Kelvin, Ratio, SceneId, Speed};